Use explicit tag for yq in image catalog build
We've been relying on the quay.io/app-sre/yq image to provide a
consistent and predictable yq for a certain step in the image catalog
build in the appsre pipeline. Unfortunately, this implicitly uses the
`latest` tag, which is a moving target. It appears as though "they"
have recently begun tagging 4.x versions as `latest`, whereas we were
relying on 3.

So with this commit, we explicitly request quay.io/app-sre/yq:3.