OSD-6853: Skip saas bundle creation and catalog build/push if image exists
Before we process the saas bundle and build/push the catalog image,
check whether that image already exists in the repository. If it does,
we can assume the saas bundle was also processed correctly, and we can
bail out of the whole catalog creation.

This serves two purposes:
- Avoids unnecessary/redundant work
- Prevents us from losing repository images due to automatic pruning
when we create a new image with the same contents & tag but a different
digest.

OSD-6853